+++ packages/classic/src/modes/ctr.ts
@@ -37,8 +37,8 @@ function createCtrTransform(cipher: BlockCipher, iv: Uint8Array): Transform {
       return output;
     },
 
-    finalize(input = new Uint8Array()) {
-      return input.length === 0 ? new Uint8Array() : this.process(input);
+    finalize(input = new Uint8Array(0)) {
+      return input.length === 0 ? new Uint8Array(0) : this.process(input);
     },
   };
 }