+++ packages/classic/src/modes/ecb.ts
@@ -25,8 +25,8 @@ function createEcbEncryptor(cipher: BlockCipher): Transform {
       return output;
     },
 
-    finalize(input = new Uint8Array()) {
-      return input.length === 0 ? new Uint8Array() : this.process(input);
+    finalize(input = new Uint8Array(0)) {
+      return input.length === 0 ? new Uint8Array(0) : this.process(input);
     },
   };
 }
@@ -44,8 +44,8 @@ function createEcbDecryptor(cipher: BlockCipher): Transform {
       return output;
     },
 
-    finalize(input = new Uint8Array()) {
-      return input.length === 0 ? new Uint8Array() : this.process(input);
+    finalize(input = new Uint8Array(0)) {
+      return input.length === 0 ? new Uint8Array(0) : this.process(input);
     },
   };
 }